+++ src/main/java/io/dnd/goyo/domain/place/entity/RegionCode.java
@@ -6,7 +6,7 @@ import jakarta.persistence.Embeddable;
 
 @Embeddable
 public record RegionCode(
-        Integer value
+        Long value
 ) {
     private static final int MIN_LENGTH = 5;
     private static final int MAX_LENGTH = 10;
@@ -16,13 +16,13 @@ public record RegionCode(
         validateCodeLength(value);
     }
 
-    private static void validateIsNotNull(Integer value) {
+    private static void validateIsNotNull(Long value) {
         if (value == null) {
             throw new BusinessException(ErrorCode.INVALID_INPUT, "행정구역 코드는 필수입니다.");
         }
     }
 
-    private static void validateCodeLength(Integer value) {
+    private static void validateCodeLength(Long value) {
         int length = String.valueOf(value).length();
         if (length < MIN_LENGTH || length > MAX_LENGTH) {
             throw new BusinessException(
